added error message handle; refactoring
