+++ src/components/dashboard/MagicLinkInfo.js
@@ -3,7 +3,7 @@ import React, { useCallback } from 'react'
 import { Platform, View } from 'react-native'
 import { t, Trans } from '@lingui/macro'
 import { fireEvent, RESENDING_MAGICLINK_SUCCESS } from '../../lib/analytics/analytics'
-import API, { getErrorMessage } from '../../lib/API/api'
+import { API, getErrorMessage } from '../../lib/API'
 import { useUserStorage } from '../../lib/wallet/GoodWalletProvider'
 import logger from '../../lib/logger/js-logger'
 import { useDialog } from '../../lib/dialog/useDialog'
