+++ src/components/signup/SignupState.js
@@ -25,7 +25,7 @@ import AuthProgressBar from '../auth/components/AuthProgressBar'
 import { navigationConfig } from '../appNavigation/navigationConfig'
 import logger from '../../lib/logger/js-logger'
 import { decorate, ExceptionCode } from '../../lib/exceptions/utils'
-import API, { getErrorMessage } from '../../lib/API/api'
+import { API, getErrorMessage } from '../../lib/API'
 import { useDialog } from '../../lib/dialog/useDialog'
 import BackButtonHandler from '../appNavigation/BackButtonHandler'
 import { showSupportDialog } from '../common/dialogs/showSupportDialog'
