+++ src/components/signup/SmsForm.js
@@ -3,7 +3,7 @@ import React, { useEffect, useRef, useState } from 'react'
 import { KeyboardAvoidingView } from 'react-native'
 import { isIOS } from '../../lib/utils/platform'
 import logger from '../../lib/logger/js-logger'
-import API, { getErrorMessage } from '../../lib/API/api'
+import { API, getErrorMessage } from '../../lib/API'
 import { getDesignRelativeHeight } from '../../lib/utils/sizes'
 import { withStyles } from '../../lib/styles'
 import SpinnerCheckMark from '../common/animations/SpinnerCheckMark'
