+++ src/lib/API/index.js
@@ -1,0 +1,2 @@
+export { default as API } from './api'
+export { getErrorMessage } from './utils'
